Switch to older nixpkgs
To avoid having to rebuild lot of things against older libxml2.

Unfortunately PHP itself still needs to be rebuilt because these
ancient nixpkgs versions did not support enabling PHP extensions
without rebuilding and we need php-tidy.

I also determined that 2.9.8 is last broken version so the fix
has to be in 2.9.9, likely the following commit:

https://gitlab.gnome.org/GNOME/libxml2/commit/35e83488505d501864826125cfe6a7950d6cba78